Added load function to Conway and moved main.py
Closes #5
